==> code/04_clean.ipynb <==
# Preprocessing text 


## Import libraries 

# Basic libs 
import pandas as pd
import string

import re

# Custom functions
from clean_text import clean_tweet

# Import files 

# Import CSV file 

df = pd.read_csv('/home/jae/hateasiancovid/processed_data/US_tweets.csv',
                 parse_dates = ['date'])


# 1,394,478 obs, 16 columns  

df.shape

# Clean text 

# Copy
df_clean = df.copy()

# Clean
df_clean['full_text'] = clean_tweet(df_clean['full_text'])


# Check 
df_clean.head()

# Export 

df_clean = df_clean.drop(columns = ['created_at', 'location'])

df_clean.to_csv("/home/jae/hateasiancovid/processed_data/cleaned_data.csv")
